Add MPP operation for measuring Pauli products (#111)
- Add decompose_mpp_operation utility method
- Move GateTarget class from pybind into main project
- Refactor OperationData to take a range of GateTarget instead of uint32_t
- Add simd_bits::intersects
- Add PointerRange::back/front
- Add MPP method to each simulator
- Add `stim.GateTarget.is_qubit_target`
- Add `stim.GateTarget.is_combiner`
- Add `stim.target_combiner`
- Fix `stim.Circuit.append_operation` not taking `stim.GateTarget` instances

Fixes https://github.com/quantumlib/Stim/issues/109

+++ src/circuit/gate_data.cc
@@ -106,7 +106,7 @@ Gate::Gate(
       extra_data_func(extra_data_func),
       flags(flags),
       arg_count(arg_count),
-      name_len(strlen(name)),
+      name_len((uint8_t)strlen(name)),
       id(gate_name_to_id(name)) {
 }
 
@@ -139,7 +139,7 @@ void GateDataMap::add_gate_alias(bool &failed, const char *alt_name, const char 
         return;
     }
     g_alt.name = alt_name;
-    g_alt.name_len = strlen(alt_name);
+    g_alt.name_len = (uint8_t)strlen(alt_name);
     g_alt.id = h_canon;
 }
